changed all names in xbox package to gamepad
