+++ hardware/gamepad_driver/launch/xbox_controller.launch.py
@@ -6,9 +6,9 @@ def generate_launch_description():
     return LaunchDescription(
         [
             Node(
-                package="xbox_controller_driver",
-                executable="xbox_controller_node",
-                name="xbox_controller_node",
+                package="gamepad_driver",
+                executable="gamepad_node",
+                name="gamepad_node",
                 output="screen",
             )
         ]
