+++ hardware/gamepad_driver/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/gamepad_driver
+[install]
+install_scripts=$base/lib/gamepad_driver
