--- hardware/xbox_controller_driver/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/xbox_controller_driver
-[install]
-install_scripts=$base/lib/xbox_controller_driver
